+++ src/Additions/Additions.php
@@ -1,33 +1,33 @@
 <?php
 /**
- * GitHub Updater Additions
+ * Git Updater Additions
  *
  * @author    Andy Fragen
  * @license   MIT
- * @link      https://github.com/afragen/github-updater-additions
- * @package   github-updater-additions
+ * @link      https://github.com/afragen/git-updater-additions
+ * @package   git-updater-additions
  */
 
-namespace Fragen\GitHub_Updater\Additions;
+namespace Fragen\Git_Updater\Additions;
 
 use Fragen\Singleton;
 
 /**
  * Class Additions
  *
- * Add repos without required headers to GitHub Updater.
+ * Add repos without required headers to Git Updater.
  * Uses JSON config data file and companion plugin.
  *
  * @uses \Fragen\Singleton
  */
 class Additions {
 	/**
-	 * Holds array of plugin/theme headers to add to GitHub Updater.
+	 * Holds array of plugin/theme headers to add to Git Updater.
 	 *
 	 * @access public
 	 * @var array
 	 */
-	public $add_to_github_updater;
+	public $add_to_git_updater;
 
 	/**
 	 * Register JSON config file.
@@ -45,16 +45,21 @@ class Additions {
 			return false;
 		}
 
+		// Bail if Git Updater not active.
+		if ( ! class_exists( '\\Fragen\\Git_Updater\\Bootstrap' ) ) {
+			return false;
+		}
+
 		$this->add_headers( $config, $repos, $type );
 
 		return true;
 	}
 
 	/**
-	 * Add GitHub Updater headers to plugins/themes via a filter hooks.
+	 * Add Git Updater headers to plugins/themes via a filter hooks.
 	 *
 	 * @access public
-	 * @uses   \Fragen\GitHub_Updater\Additions::add_to_github_updater()
+	 * @uses   \Fragen\Git_Updater\Additions::add_to_git_updater()
 	 *
 	 * @param array  $config The repo config.
 	 * @param array  $repos  The repos to pull from.
@@ -99,7 +104,7 @@ class Additions {
 					break;
 			}
 
-			$this->add_to_github_updater[ $repo['slug'] ] = array_merge( $additions[ $repo['slug'] ], $addition );
+			$this->add_to_git_updater[ $repo['slug'] ] = array_merge( $additions[ $repo['slug'] ], $addition );
 		}
 	}
 }