+++ src/Additions/Settings.php
@@ -1,14 +1,14 @@
 <?php
 /**
- * GitHub Updater Additions
+ * Git Updater Additions
  *
  * @author    Andy Fragen
  * @license   MIT
- * @link      https://github.com/afragen/github-updater-additions
- * @package   github-updater-additions
+ * @link      https://github.com/afragen/git-updater-additions
+ * @package   git-updater-additions
  */
 
-namespace Fragen\GitHub_Updater\Additions;
+namespace Fragen\Git_Updater\Additions;
 
 /**
  * Class Settings
@@ -49,7 +49,7 @@ class Settings {
 	 * Load site options.
 	 */
 	private function load_options() {
-		self::$options_additions = get_site_option( 'github_updater_additions', [] );
+		self::$options_additions = get_site_option( 'git_updater_additions', [] );
 	}
 
 	/**
@@ -57,7 +57,7 @@ class Settings {
 	 */
 	public function load_hooks() {
 		add_action(
-			'github_updater_update_settings',
+			'gu_update_settings',
 			function ( $post_data ) {
 				$this->save_settings( $post_data );
 			}
@@ -65,7 +65,7 @@ class Settings {
 		$this->add_settings_tabs();
 
 		add_filter(
-			'github_updater_add_admin_page',
+			'gu_add_admin_page',
 			function ( $tab, $action ) {
 				$this->add_admin_page( $tab, $action );
 			},
@@ -77,20 +77,20 @@ class Settings {
 	/**
 	 * Save Additions settings.
 	 *
-	 * @uses 'github_updater_update_settings' action hook
-	 * @uses 'github_updater_save_redirect' filter hook
+	 * @uses 'gu_update_settings' action hook
+	 * @uses 'gu_save_redirect' filter hook
 	 *
 	 * @param array $post_data $_POST data.
 	 */
 	public function save_settings( $post_data ) {
-		$options   = get_site_option( 'github_updater_additions', [] );
+		$options   = get_site_option( 'git_updater_additions', [] );
 		$duplicate = false;
 		$bad_input = false;
 		if ( isset( $post_data['option_page'] ) &&
-			'github_updater_additions' === $post_data['option_page']
+			'git_updater_additions' === $post_data['option_page']
 		) {
-			$new_options = isset( $post_data['github_updater_additions'] )
-				? $post_data['github_updater_additions']
+			$new_options = isset( $post_data['git_updater_additions'] )
+				? $post_data['git_updater_additions']
 				: [];
 
 			$new_options = $this->sanitize( $new_options );
@@ -110,13 +110,13 @@ class Settings {
 
 			if ( ! $duplicate && ! $bad_input ) {
 				$options = array_merge( $options, $new_options );
-				update_site_option( 'github_updater_additions', $options );
+				update_site_option( 'git_updater_additions', $options );
 			}
 
 			add_filter(
-				'github_updater_save_redirect',
+				'gu_save_redirect',
 				function ( $option_page ) {
-					return array_merge( $option_page, [ 'github_updater_additions' ] );
+					return array_merge( $option_page, [ 'git_updater_additions' ] );
 				}
 			);
 		}
@@ -126,9 +126,9 @@ class Settings {
 	 * Adds Additions tab to Settings page.
 	 */
 	public function add_settings_tabs() {
-		$install_tabs = [ 'github_updater_additions' => esc_html__( 'Additions', 'github-updater-additions' ) ];
+		$install_tabs = [ 'git_updater_additions' => esc_html__( 'Additions', 'git-updater-additions' ) ];
 		add_filter(
-			'github_updater_add_settings_tabs',
+			'gu_add_settings_tabs',
 			function ( $tabs ) use ( $install_tabs ) {
 				return array_merge( $tabs, $install_tabs );
 			},
@@ -140,7 +140,7 @@ class Settings {
 	/**
 	 * Add Settings page data via action hook.
 	 *
-	 * @uses 'github_updater_add_admin_page' action hook
+	 * @uses 'gu_add_admin_page' action hook
 	 *
 	 * @param string $tab    Tab name.
 	 * @param string $action Form action.
@@ -148,20 +148,20 @@ class Settings {
 	public function add_admin_page( $tab, $action ) {
 		$this->additions_page_init();
 
-		if ( 'github_updater_additions' === $tab ) {
+		if ( 'git_updater_additions' === $tab ) {
 			$action = add_query_arg(
 				[
-					'page' => 'github-updater',
+					'page' => 'git-updater',
 					'tab'  => $tab,
 				],
 				$action
 			);
-			( new Repo_List_Table() )->render_list_table();
+			( new Repo_List_Table( self::$options_additions ) )->render_list_table();
 			?>
 			<form class="settings" method="post" action="<?php esc_attr_e( $action ); ?>">
 				<?php
-				settings_fields( 'github_updater_additions' );
-				do_settings_sections( 'github_updater_additions' );
+				settings_fields( 'git_updater_additions' );
+				do_settings_sections( 'git_updater_additions' );
 				submit_button();
 				?>
 			</form>
@@ -174,50 +174,50 @@ class Settings {
 	 */
 	public function additions_page_init() {
 		register_setting(
-			'github_updater_additions',
-			'github_updater_additions',
+			'git_updater_additions',
+			'git_updater_additions',
 			null
 		);
 
 		add_settings_section(
-			'github_updater_additions',
+			'git_updater_additions',
 			esc_html__( 'Additions', 'github-updater' ),
 			[ $this, 'print_section_additions' ],
-			'github_updater_additions'
+			'git_updater_additions'
 		);
 
 		add_settings_field(
 			'type',
-			esc_html__( 'Repository Type', 'github-updater-additions' ),
+			esc_html__( 'Repository Type', 'git-updater-additions' ),
 			[ $this, 'callback_dropdown' ],
-			'github_updater_additions',
-			'github_updater_additions',
+			'git_updater_additions',
+			'git_updater_additions',
 			[
-				'id'      => 'github_updater_additions_type',
+				'id'      => 'git_updater_additions_type',
 				'setting' => 'type',
 			]
 		);
 
 		add_settings_field(
 			'slug',
-			esc_html__( 'Repository Slug', 'github-updater-additions' ),
+			esc_html__( 'Repository Slug', 'git-updater-additions' ),
 			[ $this, 'callback_field' ],
-			'github_updater_additions',
-			'github_updater_additions',
+			'git_updater_additions',
+			'git_updater_additions',
 			[
-				'id'      => 'github_updater_additions_slug',
+				'id'      => 'git_updater_additions_slug',
 				'setting' => 'slug',
 			]
 		);
 
 		add_settings_field(
 			'uri',
-			esc_html__( 'Repository URI', 'github-updater-additions' ),
+			esc_html__( 'Repository URI', 'git-updater-additions' ),
 			[ $this, 'callback_field' ],
-			'github_updater_additions',
-			'github_updater_additions',
+			'git_updater_additions',
+			'git_updater_additions',
 			[
-				'id'      => 'github_updater_additions_uri',
+				'id'      => 'git_updater_additions_uri',
 				'setting' => 'uri',
 			]
 		);
@@ -246,7 +246,7 @@ class Settings {
 	 */
 	public function print_section_additions() {
 		echo '<p>';
-		esc_html_e( 'If there are git repositories that do not natively support GitHub Updater you can add them here.', 'github-updater-additions' );
+		esc_html_e( 'If there are git repositories that do not natively support Git Updater you can add them here.', 'git-updater-additions' );
 		echo '</p>';
 	}
 
@@ -260,10 +260,10 @@ class Settings {
 	public function callback_field( $args ) {
 		?>
 		<label for="<?php esc_attr_e( $args['id'] ); ?>">
-			<input type="text" style="width:50%;" id="<?php esc_attr( $args['id'] ); ?>" name="github_updater_additions[<?php esc_attr_e( $args['setting'] ); ?>]" value="">
+			<input type="text" style="width:50%;" id="<?php esc_attr( $args['id'] ); ?>" name="git_updater_additions[<?php esc_attr_e( $args['setting'] ); ?>]" value="">
 			<br>
 			<span class="description">
-				<?php esc_html_e( 'Ensure proper slug for plugin or theme.', 'github-updater-additions' ); ?>
+				<?php esc_html_e( 'Ensure proper slug for plugin or theme.', 'git-updater-additions' ); ?>
 			</span>
 		</label>
 		<?php
@@ -280,7 +280,7 @@ class Settings {
 		$options['type'] = [ 'github_plugin' ];
 		?>
 		<label for="<?php esc_attr_e( $args['id'] ); ?>">
-		<select id="<?php esc_attr_e( $args['id'] ); ?>" name="github_updater_additions[<?php esc_attr_e( $args['setting'] ); ?>]">
+		<select id="<?php esc_attr_e( $args['id'] ); ?>" name="git_updater_additions[<?php esc_attr_e( $args['setting'] ); ?>]">
 		<?php
 		foreach ( self::$addition_types as $item ) {
 			printf(